Move products.py into buyersguide sub-directory
